Restructures DataStage connector in line with changes to IGC connector
